submissions/BlenderMCP-AI-AGNO/requirements.txt
agno==1.5.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
-e git+https://github.com/ahujasid/blender-mcp.git@972096e9dc0183c0ac72a5391a99c99533c7f783#egg=blender_mcp
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.0
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.60
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.42
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mcp==1.9.0
mdurl==0.1.2
numpy==1.26.2
opencv-python==4.11.0.86
orjson==3.10.18
packaging==24.2
parso==0.8.4
pefile==2024.8.26
Pillow==10.1.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2025.4
PyOpenGL==3.1.7
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dotenv==1.1.0
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
setuptools==80.7.1
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
zstandard==0.23.0
